% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheDirectory.R
\name{cacheDirectory}
\alias{cacheDirectory}
\title{Cache directory for project dependencies}
\usage{
cacheDirectory()
}
\value{
String containing the path to a cache directory.
}
\description{
Cache project dependencies so that they can be re-used across functions.
This is set to the value of the \code{FUYUKO_CACHE_DIR} environment variable if defined,
otherwise it is set to a subdirectory inside \code{\link{tempdir}}.
}
\examples{
cacheDirectory()

Sys.setenv(FUYUKO_CACHE_DIR="WHEEE")
cacheDirectory()
Sys.unsetenv("FUYUKO_CACHE_DIR")

}
\author{
Aaron Lun
}
