% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchDependencies.R
\name{fetchDependencies}
\alias{fetchDependencies}
\title{Download dependencies to disk}
\usage{
fetchDependencies(dependencies, cache)
}
\arguments{
\item{dependencies}{Data frame containing a manifest of dependencies,
typically the return value from \code{\link{scanFetchContentDeclare}}.
Each row represents a CMake project.}

\item{cache}{String containing a path to a cache directory.}
}
\value{
Character vector of length equal to the number of rows in \code{dependencies}.
Each entry contains the path to the directory (stored inside \code{cache}) for the corresponding project.
}
\description{
Fetch CMake project dependencies by cloning Git repositories or downloading their contents.
These are unpacked into a cache directory for re-use across multiple requests.
}
\author{
Aaron Lun
}
