% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryAllDependencies.R
\name{queryAllDependencies}
\alias{queryAllDependencies}
\title{Query a project for all its dependencies}
\usage{
queryAllDependencies(dir, cache = cacheDirectory())
}
\arguments{
\item{dir}{String containing a path to a CMake project directory (or one of its subdirectories).}

\item{cache}{String containing a path to a cache directory in which to cache the dependencies.}
}
\value{
A list containing:
\itemize{
\item \code{dependencies}, a data frame where each row corresponds to a CMake project that is (directly or indirectly) a dependency of \code{dir}.
This contains all columns described in \code{\link{scanFetchContentDeclare}}.
\item \code{relationships}, a data frame describing the relationships between projects in \code{dependencies}.
This contains:
\itemize{
\item \code{index}, the row index of a project in \code{dependencies}.
\item \code{parent}, the row index of the immediate parent project that requires the project referenced by \code{index}.
This is set to zero for \code{dir}.
\item \code{path}, the file inside the parent project that requires the project of \code{index}.
}
}
}
\description{
Search a CMake project for all its FetchContent-based dependencies in all its CMakeLists.txt files.
This is done recursively for all of its upstream dependencies.
}
\examples{
df <- data.frame(
    name = "scran", 
    git.repository = "https://github.com/LTLA/libscran",
    git.tag = "ae74e0345303a2d2c6e70d599d72c0e02d346fb6",
    url = NA_character_,
    url.hash = NA_character_
)

path <- fetchDependencies(df)
(out <- queryAllDependencies(path))

}
