% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeConflicts.R
\name{summarizeConflicts}
\alias{summarizeConflicts}
\title{Summarize conflicting dependencies}
\usage{
summarizeConflicts(full)
}
\arguments{
\item{full}{List produced by \code{\link{queryAllDependencies}}.}
}
\value{
List containing lists of data frames.

This includes \code{name}, a named list of data frames where each entry corresponds to a set of conflicting dependencies.
Each entry is named after a particular project's name and contains a data frame of conflicting projects with differences in properties other than the name.

The same applies for \code{git.repository}, where conflicting projects have the same Git repository but differences in other properties;
and \code{url}, for projects with the same URL but differences elsewhere.
}
\description{
Summarize conflicts between dependencies with similar but not identical properties.
}
\examples{
df <- data.frame(
    name = "scran", 
    git.repository = "https://github.com/LTLA/libscran",
    git.tag = "ae74e0345303a2d2c6e70d599d72c0e02d346fb6",
    url = NA_character_,
    url.hash = NA_character_
)

path <- fetchDependencies(df)
out <- queryAllDependencies(path)
summarizeConflicts(out)

}
